pyproject.toml
[tool.poetry]
name = "adjust_project"
version = "0.1.0"
description = ""
authors = ["NeoVic2006 <NeoVic2006@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
django-filter = "^2.4.0"
pgadmin4 = "^5.5"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
